tools: Automatically set an image name
Let's automatically set image names, not requiring the user to pass it
to our scripts.

The names chosen are basically the default ones for the Confidential
Containers project amd its CI, being:
* quay.io/confidential-containers/runtime-payload
* quay.io/confidential-containers/runtime-payload-ci

The tags, which is the part that matter for us, are being set depending
on whether we're targetting the runtime-payload registry, or the
runtime-payload-ci one.

For the former, the tag will be: `enclave-cc-${tag}-${SGX_MODE}`, where
tag is the release tag, and the SGX_MODE is either HW or SIM.

For the lkatter, the tag will be: `enclave-cc-${commit}-${SGX_MODE}`,
where the ${commit} is the hash of the current commit, and SGX_MODE is
either HW or SIM.