tools: use install to copy the installer script
enclave-cc-deploy.sh is used by the CcRuntime configuration
for enclave-cc and the current configuration expects that
script to be found under scripts/.

After f227db0 scripts/ directory is no longer created so
enclave-cc-deploy.sh is not found from the expected place.

Move away from cp to install -D to ensure all of the necessary
directory paths are created before copying the script.

Signed-off-by: Mikko Ylinen <mikko.ylinen@intel.com>